pyproject.toml
[tool.poetry]
name = "user-todo-api"
version = "0.1.0"
description = ""
authors = ["ed-machado <louchi.machado@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
fastapi = {extras = ["all"], version = "^0.112.1"}
httpx = "^0.27.0"
fastapi-cli = "^0.0.5"
sqlalchemy = "^2.0.32"
pydantic-settings = "2.3.4"
alembic = "^1.13.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev user_todo_api/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=user_todo_api -vv'
post_test = 'coverage html'

[tool.ruff]
line-length = 100
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
